PFDA-3373 trying user-token to add new site_admin (#1916)
* PFDA-3373 trying user-token to add new site_admin

* PFDA-3373 extensive logging for provision new site admin action

* PFDA-3373 User Api used for deleting org member

* PFDA-3373 added assertion for admin space deleting memberships

* PFDA-3373 removed project_leave call when removing site admin

* typo fix

+++ app/services/space_membership_service/delete.rb
@@ -1,21 +1,23 @@
 module SpaceMembershipService
   # Operation to remove a member from a space, and the associated permissions on the platform
+  # THIS IS ONLY INTENTED FOR ADMIN SPACES currently. Other space types does not remove their members via memberships, only lowers permissions to DISABLED.
   module Delete
-    # @param [DNAnexusAPI] api
+    # @param [DNAnexusAPI] admin_api - currently not used
+    # @param [DNAnexusAPI] user_api
     # @param [Space] space
     # @param [SpaceMembership] membership
     # @param [SpaceMembership] admin_membership
     def self.call(admin_api, user_api, space, membership, admin_membership)
       return false unless space
       return false unless membership
+      raise "This method is only intented for Admin Spaces at this moment!" unless space.administrator?
 
       # Should not delete last member of a Space
       raise "Cannot delete last member of space #{space.id}." if space.space_memberships.count == 1
 
       # Remove user's org and project membership from dx platform
       project_dxid = space.host_project
-      admin_api.org_remove_member(space.host_dxorg, membership.user.dxid)
-      user_api.project_leave(project_dxid)
+      user_api.org_remove_member(space.host_dxorg, membership.user.dxid)
 
       space.space_memberships.delete(membership)
       membership.destroy!
